Start to allow for script-provider parameters.
The !runscript export now merely parses the command line into argv/argv.
It then passes it wholesale to the script provider for it to do with what
it will.

This allows the script provider to implement its own private switches
(such as -m <module>) and alter the behavior of the actual script
execution.

It looks at the argv vector, stealing all switches until the first
non-switch element is encountered. It then passes the rest of the elements
to the script itself, by setting sys.argv (PySys_SetArgv) with the
remainder.

The key motivation for this is to allow switches like -m in the Python CLI
for things like pdb (the python debugger) wrap a user script.

E.g.

!runscript -m pdb myscript.py <myscript params>

just like if it were the CLI REPL:

python -m pdb myscript.py <myscript params>

Later, we'll also teach !runscript to steal the first set of params to
allow for specifying the languange (and therefore script provider) to
invoke.

I.e.

!runscript -l <lang> <script-prov-params> <script-params>

E.g.

!runscript -l py -m pdb myscript.py <myscript params>

To run Python with 'pdb' as the module to run with
(myscript.py, <myscript params>) as the sys.argv tuple.