Add support for output buffering.
By default windbg flushes the output of every callback on every Output
command. This can result in a lot of unnecessary redraws making the GUI
slow.

If a script knows it's about to dump a lot of output in a loop, it can now
call start_buffering and stop_buffering when it's done.

Now the output will be buffered in 4K chunks and flushed at the end of
script execution.