Add caching layer for GetSymbolTypeId
This function was shown to be by far the hottest path when resolving an
enum value to string in a loop, such as dumping the varheap ring buffer.

If the set of modules ever changes, we can subscribe to that event and
invalidate the cache. But this would be a pretty rare case. (Specifically,
the module caching the symbol would have to be unloaded then some other
with the same name would have to be reloaded)