Implement timed run mode for host.
This marks the first time the host steals a switch from the lower layers.

Separate host switches with provider switches with --.

If you're not passing any provider switches, no need to use --.

Examples:

!runscript hkvarheap_ringbuffer.py -d 0xac1f550000 // Valid
!runscript -t hkvarheap_ringbuffer.py -d 0xac1f550000 // Valid
!runscript -t -- -m pdb hkvarheap_ringbuffer.py -d 0xac1f550000 // Valid
!runscript -t -m pdb hkvarheap_ringbuffer.py -d 0xac1f550000  // Invalid

The last one would be interpreted as passing -m to the host, which is not
a valid host switch, but a python provider switch.