!evalstring improvements
Due to oddities in the command line parsing, !evalstring can have
different behavior depending on whether switches are passed.

E.g.

1) !evalstring "print ( 1,     3,      4, )"
2) !evalstring -- "print ( 1,     3,      4, )"
3) !evalstring -l py "print ( 1,     3,      4, )"

The bottom two will be passed down un-altered because the content enclosed
in quotes will be treated as a single token.

In the first case, the quotes are essentially eliminated by dbgeng (since
there's only one param) and so our tokenization will split it up into many
argv elements. If we were to only take the first one, we'd lose the
remainder of the string.

2 and 3 can be used as simple workarounds, but for convenience's sake, we
join all the remaining argv into a single string and pass it down to the
provider.

Also removed some stale comments.