py: Add support for virtual field access.
Add a tp_getattro callback which first calls the builtin
PyObject_GenericGetAttr routine for standard method/field resolution, and
then if that fails, attempt a lookup on a struct field.

Note: We need to reserve AttributeError for PyObject_GenericGetAttr so that
we know when an attribute was truly not found so that we don't
inadvertently continue to attempt field lookup if a legitimate error
occurred.