lua: implement getLocals/args.
This was a little tricky because I needed to prevent the parent thread
from being collected before any dependent stack frames.

For this, we store a uservalue table with the stack frame userdatum.

In that table we can hold refs to any objects we want to keep alive.

Currently we hold one ref: the parent Thread.

We store a ref to its full userdata (we can't just store a lightuserdata
pointer since that won't prevent the GC from cleaning it up)

Then when we need to get the locals, we fetch the thread from the stack
frame via its uservalue table, and off we go.