py: Improve readline/REPL support.
In DbgScriptIO_readline: translate empty strings from dbgeng into single
newlines so that Python won't consider it EOF, which causes it to bail out
of the REPL and breaks mutli-line eval support (like for loops or function
definitions spanning several lines).

Of course, this means we now need a way out: We achieve this two-fold:

1) Translate Ctrl-Z (^Z) into an empty string, which Python interprets as
EOF,
2) Support exit()/quit() gracefully by handling the SystemExit ourselves
and not exiting the debugger process.