Expose .teb for Ruby/Lua and bug fixes.
Fixed current_frame API in all providers to correctly switch to the given
thread before obtaining the current frame. Otherwise it was getting
whatever the active thread's frame was instead.

Fixed bugs where we forgot to call Switch() on CAutoSwitchThread causing
us to fail to actually do a thread switch where we needed it. (This was
overlooked due to a change in API where the ctor used to do it, but no
longer does).

Exposed .teb property in ruby/lua providers. This property was forgotten.

Updated a few sample scripts to test new/fixed functionality.