Improve command line processing.
The main motivation for this change is to make !evalstring more robust.

Previously, !evalstring would leverage the same parsing code as
!runscript, which would generate an argument vector via CommandLineToArgvW
and then re-joins the vector back into a string with a space as the
delimiter.

The problem is, CommandLineToArgvW treats double quotes (") specially [1]
which prevents common Ruby constructs like arrays of strings to be used in
their traditional syntax ["a", "b", "c"]. One could workaround it by using
single quotes, or alternate syntax like %w{a b c}, etc, but ultimately
this was not pleasant. Also, the whole back-and-forth munging of vectors
to strings was unsightly.

So I finally took the plunge and rewrote the command line handling.

First, we process the host-layer switches via strtok_s. We stop the moment
we find something that isn't a switch and leave the rest of the
commandline intact.[2]

For !evalstring, we pass that unmodified remainder to the script provider
for execution.

For !runscript, we take that remainder and call CommandLineToArgvW to
generate an argument vector. Note that only !runscript calls
CommandLineToArgvW: it does not make sense to generate an argument vector
for !evalstring: it takes one argument: the string to execute.

(Recall that previously we were generating an argument vector for both
as a side-effect of the common code path, and had to re-stitch the vector
into a single string in the !evalstring case. This hack is no longer
necessary and was removed.)

There should not be any breaking changes -- !evalstring merely supports
more syntax now.

[1] Double quotes are treated as one token, a la cmd.exe.
[2] This means double quotes are left untouched and all whitespace is
preserved (even if the languages are white-space insensitive between
tokens, it's nice to keep the user's string intact.)