Revise read_string handling.
Revised the implementation of the read_string family of APIs to better match
documentation and desired behavior. Before, we were reading beyond the NUL if
the optional 'count' parameter was provided, resulting in functions like Ruby's
'puts' to barf when it found embedded NULs in the string.

If you want embedded NULs, use the read_bytes API.

The string APIs now always terminate on the first NUL encountered. The 'count'
parameter, as documented, merely serves to limit the number of characters to
return.

As part of this change, I noticed a subtle bug in the documentation for dbgeng's
ReadMultiByteStringVirtual et al. APIs. Their MaxBytes parameters aren't
strictly honored. They might actually return more the number of bytes requested
due to an implementation detail. Will follow up with debugger folks on getting
either the docs updated, or the bug fixed. Meanwhile, I worked around it by
truncation the string myself.

Added more unit tests to improve coverage of this family of APIs.

Also introduced a first module-level DbgScript#read_string API that doesn't
require a TypedObject. This is useful for times where it would be inconvenient
to position a typed object at some random text buffer in memory.

This API is currently only implemented for narrow strings, and only in Ruby.
Expect the wide version, and other providers to follow shortly.

Also bumped the version in preparation for a future release.